Adds SchemaConfiguration class
