+++ samples/client/petstore/java/src/main/java/org/openapijsonschematools/configurations/JsonSchemaKeywordFlags.java
@@ -87,52 +87,4 @@ public record JsonSchemaKeywordFlags(
         if (unevaluatedProperties) { enabledKeywords.add("unevaluatedProperties"); }
         return enabledKeywords;
     }
-
-    @Test
-    public void testGetNoEnabledKeywords() {
-        final JsonSchemaKeywordFlags jsonSchemaKeywordFlags = new JsonSchemaKeywordFlags(
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false,
-                false
-        );
-        LinkedHashSet<String> enabledKeywords = jsonSchemaKeywordFlags.getEnabledKeywords();
-        LinkedHashSet<String> expectedEnabledKeywords = new LinkedHashSet<>();
-        Assert.assertEquals(enabledKeywords, expectedEnabledKeywords);
-    }
 }
