+++ samples/client/petstore/java/src/main/java/org/openapijsonschematools/configurations/SchemaConfiguration.java
@@ -1,0 +1,4 @@
+package org.openapijsonschematools.configurations;
+
+public record SchemaConfiguration(JsonSchemaKeywordFlags enabledKeywordFlags) {
+}
