+++ src/main/resources/java/src/main/java/org/openapitools/configurations/SchemaConfiguration.hbs
@@ -1,0 +1,4 @@
+package {{{packageName}}}.configurations;
+
+public record SchemaConfiguration(JsonSchemaKeywordFlags enabledKeywordFlags) {
+}
